requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
Pillow
google-cloud-storage
